@@ -2,8 +2,8 @@ package com.rahmat.app.samplecrudkotlin.adapter
 
 import android.content.Context
 import android.content.Intent
-import android.databinding.DataBindingUtil
-import android.support.v7.widget.RecyclerView
+import androidx.databinding.DataBindingUtil
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
