@@ -1,8 +1,8 @@
 package com.rahmat.app.samplecrudkotlin.db
 
-import android.arch.lifecycle.LiveData
-import android.arch.persistence.room.*
-import android.arch.persistence.room.OnConflictStrategy.REPLACE
+import androidx.lifecycle.LiveData
+import androidx.room.*
+import androidx.room.OnConflictStrategy.REPLACE
 import com.rahmat.app.samplecrudkotlin.entity.Student
 import io.reactivex.Flowable
 
