@@ -1,8 +1,8 @@
 package com.rahmat.app.samplecrudkotlin.db
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
-import android.arch.persistence.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
 import android.content.Context
 import com.rahmat.app.samplecrudkotlin.entity.Student
 
