@@ -1,6 +1,6 @@
 package com.rahmat.app.samplecrudkotlin.features.base
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import io.reactivex.disposables.CompositeDisposable
 
 /**
